+++ src/trx/game/gun/misc.c
@@ -67,7 +67,7 @@ static void M_SmashItem(const int16_t item_num)
         break;
 
     case O_SCION_ITEM_3:
-        Gun_HitTarget(item, nullptr, nullptr, LGT_UNARMED, item->hit_points);
+        Gun_HitTarget(item, nullptr, nullptr, item->hit_points);
         break;
 
     default:
@@ -371,26 +371,11 @@ PROJECTILE_HIT Gun_SmashItems(
 
 void Gun_HitTarget(
     ITEM *const item, const GAME_VECTOR *const start,
-    const GAME_VECTOR *const hit_pos, const LARA_GUN_TYPE weapon_type,
-    int32_t damage)
+    const GAME_VECTOR *const hit_pos, int32_t damage)
 {
-    bool make_ricochet = (g_Config.visuals.fix_texture_issues
-                          && item->object_id == O_SCION_ITEM_3)
-        || item->object_id == O_WINSTON_ARMY
-        || item->object_id == O_ASSAULT_TARGET;
-
-    if (item->object_id == O_SHIVA) {
-        const ITEM *const lara_item = Lara_GetItem();
-        const int32_t dx = item->pos.x - lara_item->pos.x;
-        const int32_t dz = item->pos.z - lara_item->pos.z;
-        const int16_t angle = DEG_180 - item->rot.y + Math_Atan(dz, dx);
-
-        if (item->current_anim_state > 1 && item->current_anim_state < 5
-            && angle > -DEG_90 && angle < DEG_90 && weapon_type != LGT_ROCKET
-            && weapon_type != LGT_GRENADE && weapon_type != LGT_HARPOON) {
-            make_ricochet = true;
-            damage = 0;
-        }
+    const bool make_ricochet = !Item_ShouldSpawnBlood(item);
+    if (item->object_id == O_SHIVA && make_ricochet) {
+        damage = 0;
     }
 
     LARA_INFO *const lara = Lara_GetLaraInfo();