+++ src/trx/game/items/utils.h
@@ -21,3 +21,5 @@ void Item_TakeDamage(ITEM *item, int16_t damage, bool hit_status);
 // * Negative values - deal damage, disable body part explosions.
 // * Zero - don't deal any damage, disable body part explosions.
 int32_t Item_Explode(int16_t item_num, int32_t mesh_bits, int16_t damage);
+
+bool Item_ShouldSpawnBlood(const ITEM *item);