+++ src/trx/game/objects/creatures/winston_army.c
@@ -39,6 +39,11 @@ typedef struct {
     bool spawn_checked;
 } M_PRIV;
 
+static bool M_ShouldSpawnBlood(const ITEM *const item)
+{
+    return false;
+}
+
 static void M_LoadPriv(ITEM *const item, SG_READ_IO *const io)
 {
     M_PRIV *const p = item->priv;
@@ -243,6 +248,7 @@ static void M_Setup(OBJECT *const obj)
     obj->priv_save_func = M_SavePriv;
     obj->control_func = M_Control;
     obj->collision_func = Object_Collision;
+    obj->should_spawn_blood_func = M_ShouldSpawnBlood;
 
     obj->hit_points = 20;
     obj->shadow_size = UNIT_SHADOW / 4;