+++ src/trx/game/objects/general/harpoon_bolt.c
@@ -1,6 +1,7 @@
 #include <trx/game/game_buf.h>
 #include <trx/game/gun/misc.h>
 #include <trx/game/gun/vars.h>
+#include <trx/game/items.h>
 #include <trx/game/lara.h>
 #include <trx/game/math.h>
 #include <trx/game/objects/vars.h>
@@ -121,11 +122,13 @@ static void M_Control_TR3(const int16_t item_num)
         }
 
         if (target_item->status == IS_ACTIVE) {
-            Spawn_BloodBath(
-                item->pos.x, item->pos.y, item->pos.z, 0, 0, item->room_num, 3);
+            if (Item_ShouldSpawnBlood(target_item)) {
+                Spawn_BloodBath(
+                    item->pos.x, item->pos.y, item->pos.z, 0, 0, item->room_num,
+                    3);
+            }
             Gun_HitTarget(
-                target_item, nullptr, nullptr, LGT_HARPOON,
-                g_Weapons[LGT_HARPOON].damage);
+                target_item, nullptr, nullptr, g_Weapons[LGT_HARPOON].damage);
             Stats_AddAmmoHits();
         }
 
@@ -273,11 +276,13 @@ static void M_Control_TR12(const int16_t item_num)
         }
 
         if (target_item->status == IS_ACTIVE) {
-            Spawn_BloodBath(
-                item->pos.x, item->pos.y, item->pos.z, 0, 0, item->room_num, 5);
+            if (Item_ShouldSpawnBlood(target_item)) {
+                Spawn_BloodBath(
+                    item->pos.x, item->pos.y, item->pos.z, 0, 0, item->room_num,
+                    5);
+            }
             Gun_HitTarget(
-                target_item, nullptr, nullptr, LGT_HARPOON,
-                g_Weapons[LGT_HARPOON].damage);
+                target_item, nullptr, nullptr, g_Weapons[LGT_HARPOON].damage);
             Stats_AddAmmoHits();
         }
         hit = true;