+++ src/trx/game/objects/general/rocket.c
@@ -156,8 +156,7 @@ static bool M_TryExplodeItem(
 
     if (target_item->status == IS_ACTIVE) {
         Gun_HitTarget(
-            target_item, nullptr, nullptr, LGT_ROCKET,
-            g_Weapons[LGT_ROCKET].damage);
+            target_item, nullptr, nullptr, g_Weapons[LGT_ROCKET].damage);
         Stats_AddAmmoHits();
 
         if (target_item->hit_points <= 0 && M_CanExplodeTarget(target_item)) {